Added PID library, fixed ROS subs

+++ microcontroller/armduino/src/Motor.cpp
@@ -1,12 +1,8 @@
-#include "Arduino.h"
+#include <Arduino.h>
 #include <Servo.h>
 #include <Stepper.h>
 #include "Motor.h"
 #include "Servo.h"
-#include "ros.h";
-#include <std_msgs/Int16.h>
-#include <std_msgs/UInt16.h>
-#include <std_msgs/Float32.h>
 
 Servo servo;
 Stepper stepper = Stepper(0,0,0);
@@ -29,7 +25,7 @@ void Motor::doPWM(float _input){
 	float input = _input;
 	servo.write(map((input*10),-10,10,0,180));
 }
-void Motor::doStepper(int16_t _input){
+void Motor::doStepper(int _input){
 	stepper.step(_input);
 }
 void Motor::doMotor(int _drive){
