+++ microcontroller/armduino/src/Motor.h
@@ -3,10 +3,7 @@
 #include "Arduino.h"
 #include <Servo.h>
 #include <Stepper.h>
-#include "ros.h";
-#include <std_msgs/Int16.h>
-#include <std_msgs/UInt16.h>
-#include <std_msgs/Float32.h>
+
 class Motor{
   public:
     Motor(int _pin, int _dirPin);
@@ -14,12 +11,13 @@ class Motor{
     Motor(Stepper _Stepper);
     Motor();
     void doMotor(int _drive);
-    void doStepper(int16_t _input);
+    void doStepper(int _input);
     void doPWM(float _input);
   private:
     int pin;
     int dirPin;
     Servo servo;
     Stepper stepper = Stepper(0,0,0);
+
 };
 #endif
